+++ src/components/common/bookmark/Bookmark.Style.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import Colors from '../../../styles/Colors';
+import { Colors } from '../../../styles';
 
 export const Bookmark = styled.div`
   display: flex;
