+++ src/components/common/chip/PointAlert.Style.ts
@@ -1,6 +1,5 @@
 import styled from 'styled-components';
-import Colors from '../../../styles/Colors';
-import FontStyles from '../../../styles/FontStyles';
+import { Colors, FontStyles } from '../../../styles';
 
 export const AddPoint = styled.div<{ $type?: 'header' | 'quiz' }>`
   ${({ $type }) => ($type === 'header' ? FontStyles.SM_Medium : FontStyles.XS_Medium)}
