+++ src/components/common/footer/Footer.Style.ts
@@ -1,6 +1,5 @@
 import styled from 'styled-components';
-import Colors from '../../../styles/Colors';
-import FontStyles from '../../../styles/FontStyles';
+import { Colors, FontStyles } from '../../../styles';
 
 export const Footer = styled.footer`
   ${FontStyles.SM_Medium}
