+++ src/components/common/tabBar/TabBar.Style.ts
@@ -1,6 +1,5 @@
 import styled from 'styled-components';
-import FontStyles from '../../../styles/FontStyles';
-import Colors from '../../../styles/Colors';
+import { Colors, FontStyles } from '../../../styles';
 
 export const TabBar = styled.div`
   display: flex;
