+++ src/components/newsDetail/emojiCount/EmojiCount.Style.ts
@@ -1,6 +1,5 @@
 import styled from 'styled-components';
-import Colors from '../../../styles/Colors';
-import FontStyles from '../../../styles/FontStyles';
+import { Colors, FontStyles } from '../../../styles';
 
 export const Container = styled.div<{ $isActive: boolean }>`
   ${FontStyles.SM_Medium}
