+++ src/styles/index.ts
@@ -1,0 +1,2 @@
+export { default as Colors } from './Colors';
+export { default as FontStyles } from './FontStyles';
